解决Mark过多时的抖动和跳出问题，完成基本功能

+++ app/src/main/java/com/example/liuhui/photonote/Mark.java
@@ -9,16 +9,16 @@ import org.litepal.crud.DataSupport;
 public class Mark extends DataSupport {
 
     /* mark所处的横坐标 */
-    private double x;
+    private float x;
     /* mark所处的纵坐标 */
-    private double y;
+    private float y;
     /* mark所属的笔记的id */
     private long noteId;
     /* mark所携带的信息 */
     private String mess;
 
     /* 构造器 */
-    public Mark(double x, double y, long noteId, String mess) {
+    public Mark(float x, float y, long noteId, String mess) {
         setX(x);
         setY(y);
         setNoteId(noteId);
@@ -30,19 +30,19 @@ public class Mark extends DataSupport {
 
     /* getter and setter */
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
